requirements.txt
amqp==2.5.2
APScheduler==3.6.3
asgiref==3.3.1
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
autopep8==1.5
Babel==2.6.0
beautifulsoup4==4.8.0
billiard==3.6.3.0
celery==4.4.2
certifi==2019.3.9
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
Click==7.0
configparser==4.0.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==2.8
cssutils==1.0.2
daphne==2.5.0
Django==2.2.1
django-audit-log==0.7.0
django-cleanup==4.0.0
django-cors-headers==3.0.1
django-cors-middleware==1.4.0
django-discover-runner==1.0
django-fcm==0.1.1
django-filter==2.1.0
django-inlinecss==0.3.0
django-oauth-toolkit==1.2.0
django-phone-field==1.7.1
django-phonenumber-field==2.4.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-sendgrid-v5==0.8.1
django-taggit==1.2.0
django-taggit-serializer==0.1.7
django-user-tracking==0.4.11
django-uuidfield==0.5.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-ignore-slash-middleware==0.0.1
drf-tracking==1.5.0
fcm-django==0.2.21
future==0.18.2
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.6.0
incremental==17.5.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
kombu==4.6.8
lazy-object-proxy==1.4.3
livereload==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.0.4
more-itertools==8.2.0
msgpack-python==0.5.6
oauthlib==3.0.2
oci==2.14.1
openapi-codec==1.3.2
packaging==20.3
phonenumbers==8.10.12
Pillow==7.0.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyfcm==1.4.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pynliner==0.8.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
python-dateutil==2.8.1
python-decouple==3.1
python-http-client==3.2.7
python-memcached==1.59
pytz==2019.3
PyYAML==5.1
redis==3.2.0
requests==2.21.0
requests-toolbelt==0.9.1
sendgrid==6.2.1
service-identity==18.1.0
simplejson==3.17.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.1
tornado==6.0.2
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.1.0
zope.interface==5.1.0
codecov
sentry-sdk==0.19.5
